requirements.txt
langchain
torch
accelerate
sentence-transformers
streamlit-chat
streamlit
faiss-cpu
tiktoken
huggingface-hub
pypdf
llama-cpp-python
nltk
PyPDF2
python-dotenv
paramiko
langdetect
googletrans==4.0.0-rc1
transformers
